requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# certifi==2021.5.30
chromedriver-autoinstaller==0.6.4
Custom-Tooltip==1.4.1
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pillow==12.0.0
python-dotenv==1.1.1
redis==5.2.1
requests==2.25.1
selenium==4.37.0
spotipy==2.25.1
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==1.26.6
websocket-client==1.9.0
ytmusicapi==1.11.1
